[sprint_2] рефакторинг тестового компонента paragraph
