+++ src/app/components/paragraph/paragraph.ts
@@ -1,0 +1,12 @@
+import { Block } from '../../modules/block';
+import template from './paragraph.hbs';
+
+export class Paragraph extends Block {
+  constructor(props: any) {
+    super(props);
+  }
+
+  render(): DocumentFragment {
+    return this.compile(template, { ...this.props });
+  }
+}
